server/parser-service/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
aiofiles==23.2.1

# Document Parsing
pymupdf==1.23.8
python-docx==1.1.0
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9
beautifulsoup4==4.12.2
lxml==4.9.3

# Image Processing & OCR
pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78

# Export Formats
markdown==3.5.1

# Background Tasks
redis==5.0.1
celery==5.3.4

# NEW: Additional Format Support
striprtf==0.0.26           # RTF parsing
odfpy==1.4.1               # ODT parsing
chardet==5.2.0             # Character encoding detection
python-magic==0.4.27       # File type detection

# NEW: NLP & Analysis
langdetect==1.0.9          # Language detection
python-dateutil==2.8.2     # Date parsing
phonenumbers==8.13.25      # Phone number parsing
validators==0.22.0         # URL/email validation

# NEW: Advanced NLP (Semantic Analysis)
scikit-learn==1.3.2        # Machine learning (TF-IDF, classification)
gensim==4.3.2              # Topic modeling (LDA)
nltk==3.8.1                # Natural language toolkit
pymorphy2==0.9.1           # Russian morphology
spacy==3.7.2               # Advanced NER (optional, heavy)

# NEW: Data Cleaning
ftfy==6.1.3                # Text cleaning
python-Levenshtein==0.23.0 # String similarity

# NEW: Archive Support
py7zr==0.20.8              # 7Z archives
rarfile==4.1               # RAR archives
zipfile38==0.0.3           # Extended ZIP support

# NEW: Security & Auth
PyJWT==2.8.0               # JWT tokens
passlib==1.7.4             # Password hashing
python-jose[cryptography]==3.3.0  # JWT handling
bcrypt==4.1.2              # Password encryption

# NEW: Monitoring & Performance
prometheus-client==0.19.0  # Metrics
psutil==5.9.6              # System monitoring
orjson==3.9.10             # Fast JSON

# NEW: Database (for metadata & caching)
sqlalchemy==2.0.23         # ORM
alembic==1.13.0            # Migrations
asyncpg==0.29.0            # PostgreSQL async driver

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0
